+++ src/hostname__/index.d.ts
@@ -1,4 +1,9 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const hostname__:Be<WritableAtom_<string|undefined>>
-export { hostname__ as hostname$_ }
+import type { ReadableAtom_ } from '@ctx-core/nanostores'
+import type { Be, Ctx } from '@ctx-core/object'
+export declare const hostname__:Be<ReadableAtom_<string>>
+export {
+	hostname__ as hostname$_,
+	hostname__ as window_location_hostname__,
+	hostname__ as window_location_hostname$_
+}
+export declare function hostname_(ctx:Ctx):string