+++ src/window_location__/index.d.ts
@@ -1,7 +1,6 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const window_location__:Be<window_location__T>
-export interface window_location__T extends WritableAtom_<Location|undefined> {
-	window_location__reset:()=>void;
-}
+import type { ReadableAtom_ } from '@ctx-core/nanostores'
+import type { Be, Ctx } from '@ctx-core/object'
+export declare const window_location__:Be<ReadableAtom_<Location>>
+export declare function window_location_(ctx:Ctx):Location
 export { window_location__ as window_location$_ }
+export declare function window_location__reset(ctx:Ctx):void