+++ src/window_location_hostname__/index.d.ts
@@ -1,4 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const window_location_hostname__:Be<ReadableAtom_<string|undefined>>
-export { window_location_hostname__ as window_location_hostname$_ }