+++ src/window_location_pathname__/index.d.ts
@@ -1,5 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const window_location_pathname__:Be<ReadableAtom_<window_location_pathname_T>>
-export declare type window_location_pathname_T = string|undefined
-export { window_location_pathname__ as window_location_pathname$_ }